Modified the generation of stochastic sources in a way that gives the same result independent of the number of cpus used. The Base::Z2 randum number generator is flushed on creation a couple of times by default in order to avoid patches of same sign at the beginning of a random sequence.
Moreover I started implementation of an iterator over single timeslices, which is going to involve some additional work on the architecture dependent code.

+++ src/L0/Base/Z2.h
@@ -10,8 +10,8 @@ namespace Base
   {
     static Z2 *s_instance;
 
-    static uint64_t const s_mask64 = 0x8000000000000000;
-    static uint64_t const s_maskPoly = 0x56C9E91ACA649B2C; // More dense polynomial
+    static uint64_t const s_mask64 = uint64_t(0x8000000000000000);
+    static uint64_t const s_maskPoly = uint64_t(0x56C9E91ACA649B2C); // More dense polynomial
 
     uint64_t  d_state;
